Week2-Task8: Create Coverter component
