+++ src/App.js
@@ -1,68 +1,28 @@
 import React, { useState, useEffect } from 'react';
-import Amount from './components/Amount';
+import Converter from './components/Converter';
 import SwapThemeButton from './components/SwapThemeButton';
 import { ThemeProvider } from './components/ThemeContext';
 import './App.css';
 
-const calcExchangeRate = () => {
-	return Math.random() * 10000;
-}
-
 const App = () => {
-
-  const [state, setState] = useState({eur: 0, btc: 0});
-  const [exchangeRate, setExchangeRate] = useState(calcExchangeRate);
-  const {eur, btc} = state;
-
-  useEffect(() => {
-		let timer = setTimeout(()=>{setExchangeRate(0)}, 5000);
-
-		return () => {
-			clearTimeout(timer);
-		}
-	},[state]);
-
-  const onChange = (value) => {
-
-    let calcEur = value;
-    let calcBtc = value * parseInt(exchangeRate, 10);
-
-    setState({
-      eur: calcEur,
-      btc: calcBtc,
-    });
-
-    if (exchangeRate === 0) {
-			setExchangeRate(calcExchangeRate);
-		}
-  }
 
   return (
     <ThemeProvider>
     <div className="App">
       <div className="container">
         <h2 className="center">Zer0 to Hero</h2>
-        <form>
-          <Amount 
-            id="eur"
-            name="eur"
-            label="Euro" 
-            className=""
-            value={eur}
-            onChange={onChange}
-            readonly={false}/> 
-
-          <Amount 
-            id="btc"
-            name="btc"
-            label="BTC" 
-            className=""
-            value={btc}
-            readonly={true}/> 
-
-          <SwapThemeButton />
-        </form>
-    </div>
+        <Converter 
+          cryptoName={"btc"}
+          exchangeRate={995}
+          className=""
+        />
+        <Converter 
+          cryptoName={"eth"}
+          exchangeRate={1.2}
+          className=""
+        />
+        <SwapThemeButton />
+      </div>
     </div>
     </ThemeProvider>
   );
